labs/voice-assistant/requirements.txt
openai-whisper
torch
fastapi
uvicorn[standard]
pyttsx3
pydantic
numpy
websockets
python-multipart
